Added tests for `threads.js`
Refactored out thread management to `threadsManager.js` to keep the
command logic separate from the storage logic.

Note: I think this has found a bug. Thread IDs are added with trailing
spaces, but they aren't searched or removed with the extra space.

I haven't fixed in this PR as I think it needs some handling for both
states as they may end up being mixed in storage. This opportunity for
scope will be tracked in a new issue.